requirements_tensorflow_gpu.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==20.2.0
blis==0.7.5
cachetools==4.1.1
catalogue==2.0.6
certifi==2020.6.20
chardet==3.0.4
click==8.0.4
cymem==2.0.6
dill==0.3.2
future==0.18.2
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Jinja2==3.0.3
joblib==0.16.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
Markdown==3.2.2
MarkupSafe==2.0.1
murmurhash==1.0.6
nltk==3.6.7
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.4
pathy==0.6.1
preshed==3.0.6
promise==2.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
setuptools==59.5.0
six==1.15.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-metadata==0.24.0
termcolor==1.1.0
thinc==8.0.13
threadpoolctl==2.1.0
tqdm==4.49.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.25.10
wasabi==0.9.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zmq==0.0.0